cmake config files
Hide implementation inside the library

make it a factory
